pyproject.toml
[project]
name = "hyperopt-llm-project"
version = "0.1.0"
description = "Hyperparameter Optimization with LLMs, BO, MLflow"
requires-python = ">=3.10"

dependencies = [
    "transformers>=4.35.0",
    "huggingface_hub>=0.19.0",
    "accelerate>=0.24.0",

    "numpy>=1.26.4",            # fixed syntax

    "pandas>=2.1.0",
    "scikit-learn>=1.3",
    "torch>=2.1.0",
    "matplotlib>=3.8",
    "seaborn>=0.13.0",
    "plotly>=5.18.0",
     
    "optuna>=3.4.0",
    "scikit-optimize>=0.10.1",

    "mlflow>=2.10.0",

    "langchain==0.1.15",        # pinned for consistency
    "streamlit>=1.30.0",

    "pyyaml>=6.0",
    # ✅ Added
    "ollama>=0.1.7",
    "statsmodels>=0.14.0",
    "nbformat",
    "kaleido"
]

[tool.uv]
dev-dependencies = [
    "pytest",
    "ruff",
    "black"
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"
